Replace swift-telegram-bot with generated OpenAPI client
Remove the nerzh/swift-telegram-bot dependency (and its transitive
swift-custom-logger, swift-regular-expression pins) and generate the
Telegram Bot API types and client from an OpenAPI 3.0 spec (Bot API 10.3)
with apple/swift-openapi-generator.

- Add TelegramBotAPI target with committed generated code; consumers only
  need swift-openapi-runtime + transport, the generator itself is pinned
  in Tooling/APIGen outside the main package graph
- Add spec toolchain: openapi/ (spec, patched spec, generator config),
  Scripts/patch-openapi.py (inline 554 single-ref allOf wrappers, hoist
  shared ChatId/ReplyMarkup schemas, InputFile -> binary),
  Scripts/generate-api.sh
- Add Teleroute-owned Telegram layer: TelegramBotClient (typed wrappers
  unwrapping the {ok, result} envelope, full API via bot.api),
  TelegramLongPollingConnection (offset tracking, jittered backoff),
  TelegramAPIError (decoded from undocumented non-200 responses),
  token-bucket rate limit middleware (30 req/s, getUpdates exempt),
  prefix-free vocabulary (Update, Message, ChatId, ParseMode, ...)
- TelerouteBot now encapsulates the client: init(token:) builds the
  AsyncHTTPClient transport internally; drop the nerzh dispatcher
  indirection and the public attach() step
- Rewrite TelerouteTestSupport around a fake ClientTransport dispatching
  on operationID; add tests for error mapping, polling, and rate limiting

Breaking: TG-prefixed types are gone, updateId/messageId are Int64,
TelerouteBot is constructed from a token instead of a TGBot.

+++ Sources/Teleroute/Routing/TelerouteEvents.swift
@@ -48,7 +48,7 @@ public struct TelerouteEvent: Equatable, Sendable {
     /// Matched route name or callback pattern when available.
     public let routeName: String?
     /// Telegram update identifier.
-    public let updateId: Int
+    public let updateId: Int64
     /// Resolved chat identifier, if present in the update.
     public let chatId: Int64?
     /// Resolved user identifier, if present in the update.
@@ -67,7 +67,7 @@ public struct TelerouteEvent: Equatable, Sendable {
         kind: Kind,
         routeKind: RouteKind,
         routeName: String? = nil,
-        updateId: Int,
+        updateId: Int64,
         chatId: Int64?,
         userId: Int64?,
         startedAt: ContinuousClock.Instant = ContinuousClock().now,
